Move pkg/subctl/operator/common/embeddedyamls under pkg
