+++ pkg/embeddedyamls/.gitignore
