+++ pkg/embeddedyamls/generate.go
@@ -17,4 +17,4 @@ limitations under the License.
 */
 package embeddedyamls
 
-//go:generate go run generators/yamls2go.go ../../../../../ .
+//go:generate go run generators/yamls2go.go ../../ .
