+++ pkg/embeddedyamls/generators/yamls2go.go
