+++ pkg/subctl/operator/submarinerop/crds/ensure.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/submariner-io/submariner-operator/pkg/crd"
-	"github.com/submariner-io/submariner-operator/pkg/subctl/operator/common/embeddedyamls"
+	"github.com/submariner-io/submariner-operator/pkg/embeddedyamls"
 )
 
 // Ensure functions updates or installs the operator CRDs in the cluster.
