+++ pkg/subctl/operator/submarinerop/serviceaccount/ensure.go
@@ -20,7 +20,7 @@ package serviceaccount
 
 import (
 	"github.com/pkg/errors"
-	"github.com/submariner-io/submariner-operator/pkg/subctl/operator/common/embeddedyamls"
+	"github.com/submariner-io/submariner-operator/pkg/embeddedyamls"
 	"github.com/submariner-io/submariner-operator/pkg/subctl/operator/common/serviceaccount"
 	"k8s.io/client-go/kubernetes"
 )
